Merge with PostgreSQL, up to the point where the REL9_4_STABLE branch was
created, and 9.5 development started on the PostgreSQL master branch. The exact
upto commit is ab76208e3df6841b3770edeece57d0f048392237 which falls in between
9.4beta1 and 9.4beta2.

Notable upstream changes, from a GPDB point of view:

* ALTER SYSTEM. New SQL command for changing postgresql.conf configuration file
  entries. Currently, just changes the value on QD. Not been made to work in
  GPDB yet for the full cluster including primaries and mirrors. It can be more
  convenient than the traditional method of manually editing the postgresql.conf
  file and is alternative to GPDB `gpconfig` tool. Thought since this command
  acts directly on the file system and cannot be rolled back, it is not allowed
  inside a transaction block or function. So, needs to be evaluated how to
  correctly implement in GPDB.

* Add support for logical decoding of WAL data, to allow database changes to be
  streamed out in a customizable format. Nothing interesting has been done to
  make it work for GPDB yet. Discussion was initiated to figure out
  possibilities and general direction to leverage it in GPDB
  https://groups.google.com/a/greenplum.org/d/msg/gpdb-dev/rs7Dozj8_kk/GYWCRpYhBQAJ.

* Allow background worker processes to be dynamically registered, started and
  terminated.

* Allow dynamic allocation of shared memory segments.

* SnapshotNow and HeapTupleSatisfiesNow() goes away. All existing uses have been
  switched to more appropriate snapshot types. Catalog scans now use MVCC
  snapshots. In GPDB the segments (QE), the "regular" MVCC snapshot is a
  segment-local snapshot.

* Replication Slots. Allows to coordinate activity on streaming standbys with
  the node they are streaming from. Replication slots allow preservation of
  resources like WAL files on the primary until they are no longer needed by
  standby servers. TODO: GPDB taking advantage for replication slots to manage
  the WAL files for master and primaries. Need to find out how to expose this
  feature in MPP setting, as might not be easy for administrators to manage a
  replication slot for each primary and mirror pair.

* pg_lsn data type. 9.4 added to represent a WAL log sequence number
  (LSN). TODO: GPDB could leverage this new datatype in places where string
  based operations were performed for LSN.

* pg_basebackup allows relocating tablespaces in the backup copy. This option
  needs to be leveraged by tools like `gprecoverseg` and `gpinitstandby` calling
  `pg_basebackup` internally in GPDB to properly handle tablespaces.

* Configure flag that appends custom text to the PG_VERSION string. TODO: GPDB
  should drop the `BUILD_NUMBER` and `generate_build_number()` hack and instead
  use this to add the custom part to VERSION string.

* Allow multiple backends to insert into WAL buffers concurrently. This improves
  parallel write performance. Very helpful improvement for GPDB given append
  optimized tables also started writing WAL.

* Huge page support for shared memory. huge_pages parameter allows using huge
  memory pages on Linux. Would be great to explore how can be leveraged in GPDB
  and recommendations around it.

+++ src/backend/access/hash/hashinsert.c
@@ -3,7 +3,7 @@
  * hashinsert.c
  *	  Item insertion in hash tables for Postgres.
  *
- * Portions Copyright (c) 1996-2013, PostgreSQL Global Development Group
+ * Portions Copyright (c) 1996-2014, PostgreSQL Global Development Group
  * Portions Copyright (c) 1994, Regents of the University of California
  *
  *
@@ -65,9 +65,8 @@ _hash_doinsert(Relation rel, IndexTuple itup)
 	if (itemsz > HashMaxItemSize((Page) metap))
 		ereport(ERROR,
 				(errcode(ERRCODE_PROGRAM_LIMIT_EXCEEDED),
-				 errmsg("index row size %lu exceeds hash maximum %lu",
-						(unsigned long) itemsz,
-						(unsigned long) HashMaxItemSize((Page) metap)),
+				 errmsg("index row size %zu exceeds hash maximum %zu",
+						itemsz, HashMaxItemSize((Page) metap)),
 			errhint("Values larger than a buffer page cannot be indexed.")));
 
 	/*
@@ -90,7 +89,7 @@ _hash_doinsert(Relation rel, IndexTuple itup)
 
 		/*
 		 * If the previous iteration of this loop locked what is still the
-		 * correct target bucket, we are done.	Otherwise, drop any old lock
+		 * correct target bucket, we are done.  Otherwise, drop any old lock
 		 * and lock what now appears to be the correct bucket.
 		 */
 		if (retry)
