+++ src/backend/cdb/dispatcher/cdbdispatchresult.c
@@ -665,7 +665,7 @@ cdbdisp_sumRejectedRows(CdbDispatchResults *results)
 	CdbDispatchResult *dispatchResult;
 	CdbDispatchResult *resultEnd = cdbdisp_resultEnd(results, -1);
 	PGresult   *pgresult;
-	int			totalRejected = 0;
+	uint64		totalRejected = 0;
 
 	for (dispatchResult = cdbdisp_resultBegin(results, -1);
 		 dispatchResult < resultEnd; ++dispatchResult)
